--- react-components/src/components/Card/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Card'
