+++ react-components/src/components/CardHome/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardHome'
